Application config for local config & Kubernetes-API
Using reasonable defaults, so we don't have to change too much
when running in different scenarios.

The aggregate discovery mode isn't currently falling back
from the kubernetes to the config mode, as any exceptions will
blow it up. So we're just using externally injected config
to tell it to use the kubernetes bootstrap approach when we
launch inside kubernetes.

Include hostname was picking up the HOSTNAME environment variable
from GKE, which was resolving to the pod name, which is different
than the fully-qualified hostname.

This was causing the election to reject the join requests coming
over the network, as they would not consider the self-node to be
part of that group.